:sparkles: 관리자 역사 페이지 삭제 기능 구현
